First iteration of script using Gtk
I think a few updates need to be made, like changing the code searching to the one Rena has to work on the GC system and most likely will need to add a scroll bar for the file list. Also need to figure out why label cannot be hidden for selecting from file list until selecting "Okay." I'm just uploading this for reference if anyone wants to see it tomorrow. I'll be making updates on it throughout the rest of the week when I find time.